feat: naive implementation of an entropy lookup.
The algo differs from the ruby version.
The ruby version scraped slices from the side with a hardcoded (but
configurable) width and then height.

Here, we scan the image by looking for square regions of width and
height that have the highest entropy.

This (probably) performs worse, but the result is better. And we can
improve the performace by juggling references, cloning less and such.